Migrate Channels and transition Ammenity Model
Making all models' attributes nullable becomes a must, is just the first step of the cruzade

+++ webapp/src/main/java/ar/edu/itba/paw/webapp/controller/AdminController.java
@@ -205,7 +205,7 @@ public class AdminController {
 
     @RequestMapping(value = "/delete-amenity/{id}", method = RequestMethod.GET)
     public ModelAndView deleteAmenity(
-            @PathVariable(value = "id") int amenityId
+            @PathVariable(value = "id") Long amenityId
     ) {
         ModelAndView mav = new ModelAndView("redirect:/admin/amenities");
         as.deleteAmenity(amenityId); //JOAAAAAAAAAAAAAAAAAAAEWQDSAFEDAs
